separate more opera code from main config
