+++ src/dolphin/workflows/s1_disp.py
@@ -9,10 +9,11 @@ from pprint import pformat
 from dolphin import __version__
 from dolphin._background import DummyProcessPoolExecutor
 from dolphin._log import get_log, log_runtime
+from dolphin.opera_utils import group_by_burst
 from dolphin.utils import get_max_memory_usage, set_num_threads
 
 from . import stitch_and_unwrap, wrapped_phase
-from ._utils import _create_burst_cfg, _remove_dir_if_empty, group_by_burst
+from ._utils import _create_burst_cfg, _remove_dir_if_empty
 from .config import Workflow
